+++ service/controller/clusterapi/resource.go
@@ -16,6 +16,7 @@ import (
 	"github.com/giantswarm/prometheus-meta-operator/v2/pkg/cluster"
 	"github.com/giantswarm/prometheus-meta-operator/v2/pkg/organization"
 	"github.com/giantswarm/prometheus-meta-operator/v2/pkg/password"
+	"github.com/giantswarm/prometheus-meta-operator/v2/pkg/prometheus/agent"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/alerting/alertmanagerwiring"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/alerting/heartbeat"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/alerting/heartbeatwebhookconfig"
@@ -68,7 +69,7 @@ type Config struct {
 	PrometheusImageRepository    string
 	PrometheusVersion            string
 
-	PrometheusAgentShardingStrategy remotewriteconfig.PrometheusAgentShardingStrategy
+	ShardingStrategy agent.ShardingStrategy
 
 	RestrictedAccessEnabled bool
 	WhitelistedSubnets      string
@@ -198,7 +199,7 @@ func New(config Config) ([]resource.Interface, error) {
 			Region:       config.Region,
 			Version:      config.PrometheusVersion,
 
-			PrometheusAgentShardingStrategy: config.PrometheusAgentShardingStrategy,
+			ShardingStrategy: config.ShardingStrategy,
 		}
 
 		remoteWriteConfigResource, err = remotewriteconfig.New(c)